If ny facility is added in room then it cannot be removed

+++ Admin/ajax/features_facilities.php
@@ -102,16 +102,23 @@ if (isset($_POST['rem_facility'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_facility']];
 
-    $pre_q = "SELECT* FROM `facilities` WHERE `id`=?";
-    $res = select($pre_q, $values, 'i');
-    $img = mysqli_fetch_assoc($res);
+    $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?', [$frm_data['rem_facility']], 'i');
 
-    if (deleteImage($img['icon'], FACILITIES_FOLDER)) {
-        $q = "DELETE FROM `facilities` WHERE `id`=?";
-        $res = delete($q, $values, 'i');
-        echo $res;
+    if (mysqli_num_rows($check_q) == 0) {
+        $pre_q = "SELECT* FROM `facilities` WHERE `id`=?";
+        $res = select($pre_q, $values, 'i');
+        $img = mysqli_fetch_assoc($res);
+
+        if (deleteImage($img['icon'], FACILITIES_FOLDER)) {
+            $q = "DELETE FROM `facilities` WHERE `id`=?";
+            $res = delete($q, $values, 'i');
+            echo $res;
+        } else {
+            echo 0;
+        }
     } else {
-        echo 0;
+        echo 'room_added';
     }
+
 }
 ?>
